Move Shearwater DB import into its own file
